BUG#: 6037 TITLE: PEP 286 Privilege Separation DESCRIPTION: Implement privilege separation for the CIM Server.  This allows the CIM Server to run without privilege, using a separate small process to perform privileged operations as necessary.

+++ src/Pegasus/Security/Authentication/SecureLocalAuthenticator.cpp
@@ -34,6 +34,8 @@
 
 #include <Pegasus/Common/FileSystem.h>
 #include <Pegasus/Common/Tracer.h>
+#include <Pegasus/Common/Executor.h>
+#include <Executor/Strlcpy.h>
 
 #include "LocalAuthFile.h"
 #include "SecureLocalAuthenticator.h"
@@ -72,44 +74,57 @@ SecureLocalAuthenticator::~SecureLocalAuthenticator()
 //
 // Does local authentication
 //
-Boolean SecureLocalAuthenticator::authenticate
-(
+Boolean SecureLocalAuthenticator::authenticate(
    const String& filePath, 
    const String& secretReceived, 
-   const String& secretKept
-)
+   const String& secretKept)
 {
     PEG_METHOD_ENTER(TRC_AUTHENTICATION,
         "SecureLocalAuthenticator::authenticate()");
 
     Boolean authenticated = false;
 
+    // Use executor, if present.
 
-    if ((!String::equal(secretReceived, String::EMPTY)) &&
-        (!String::equal(secretKept, String::EMPTY)))
+    if (Executor::detectExecutor() == 0)
     {
-        if (String::equal(secretKept, secretReceived))
+        if (!String::equal(secretKept, String::EMPTY) &&
+            String::equal(secretKept, secretReceived))
+        {
+            authenticated = true;
+        }
+        else if (Executor::authenticateLocal(
+            (const char*)filePath.getCString(),
+            (const char*)secretReceived.getCString()) == 0)
         {
             authenticated = true;
         }
     }
-
-    //
-    // remove the auth file created for this user request
-    //
-    if (filePath.size())
+    else
     {
-        if (FileSystem::exists(filePath))
+        // Check secret.
+
+        if (!String::equal(secretKept, String::EMPTY) &&
+            String::equal(secretKept, secretReceived))
         {
-            FileSystem::removeFile(filePath);
+            authenticated = true;
+        }
+
+        // Remove the auth file created for this user request
+
+        if (filePath.size())
+        {
+            if (FileSystem::exists(filePath))
+            {
+                FileSystem::removeFile(filePath);
+            }
         }
     }
 
     PEG_METHOD_EXIT();
 
-    return (authenticated);
+    return authenticated;
 }
-
 
 Boolean SecureLocalAuthenticator::validateUser (const String& userName)
 {
@@ -131,8 +146,9 @@ Boolean SecureLocalAuthenticator::validateUser (const String& userName)
 // Create authentication response header
 //
 String SecureLocalAuthenticator::getAuthResponseHeader(
-    const String& authType, 
-    const String& userName, 
+    const String& authType,
+    const String& userName,
+    String& filePath,
     String& secret)
 {
     PEG_METHOD_ENTER(TRC_AUTHENTICATION,
@@ -142,27 +158,42 @@ String SecureLocalAuthenticator::getAuthResponseHeader(
     responseHeader.append(authType);
     responseHeader.append(" \"");
 
-    //
-    // create a file using user name and write a random number in it.
-    //
-    LocalAuthFile localAuthFile(userName);
-    String filePath  = localAuthFile.create();
+    // Use executor, if present.
 
-    //
-    // get the secret string
-    //
-    secret = localAuthFile.getSecretString();
+    if (Executor::detectExecutor() == 0)
+    {
+        char filePathBuffer[EXECUTOR_BUFFER_SIZE];
 
-    // 
-    // build response header with file path and challenge string.
-    //
-    responseHeader.append(filePath);
-    responseHeader.append("\"");
+        if (Executor::challengeLocal(
+                userName.getCString(), filePathBuffer) != 0)
+        {
+            throw CannotOpenFile(filePathBuffer);
+        }
+        filePath = filePathBuffer;
+        secret.clear();
+
+        responseHeader.append(filePath);
+        responseHeader.append("\"");
+    }
+    else
+    {
+        // create a file using user name and write a random number in it.
+        LocalAuthFile localAuthFile(userName);
+        filePath = localAuthFile.create();
+
+        //
+        // get the secret string
+        //
+        secret = localAuthFile.getSecretString();
+
+        // build response header with file path and challenge string.
+        responseHeader.append(filePath);
+        responseHeader.append("\"");
+    }
 
     PEG_METHOD_EXIT();
 
-    return (responseHeader);
+    return responseHeader;
 }
-
 
 PEGASUS_NAMESPACE_END
